feat: 헤더 알림 부분 스타일과, 상태관리 페이지 분리
